+++ AssessmentDatabase/Utils/AssessmentsUtils.swift
@@ -21,8 +21,8 @@ public class AssessmentsUtils: EntityUtils {
     }
     
     public func copyFields(from item: AssessmentFields, to entity: Assessment) {
-        entity.sid = Int64(item.sid)
-        entity.schoolId = Int64(item.schoolId)
+        entity.sid = item.sid
+        entity.schoolId = item.schoolId
         entity.date = item.date as NSDate
         entity.isSynced = item.isSynced
     }
@@ -88,7 +88,7 @@ public class AssessmentsUtils: EntityUtils {
     }
     
     private func set(
-        instructorSid: Int64,
+        instructorSid: Int,
         of assessment: Assessment,
         in context: NSManagedObjectContext) throws
     {
